Run mod tidy before mage
It's useful to run "go mod tidy" to download missing packages and update
go.sum, but that doesn't work if it's run from the magefile because the
magefile needs an up-to-date module cache and go.sum before it builds
and runs itself. It seems more reliable to run that command in advance
instead.