Add test-building and -running target helpers
Multiple other projects use these same commands to build the tests based
on which files changed, so it makes sense to refactor them into this
module instead. The difficult part is that this module doesn't have any
tests of its own, so it can't use (and therefore test) these functions
itself.