Make Stringer work in main directory
Confusion over what's a directory and what's a package caused Stringer
to fail to find types when they were defined in the main module
directory. We avoid that problem now by making directory paths be
abaolute.

Fixes #2.

This also adds an example project for Stringer that demonstrates the
problem and gets built as part of this project's tests.