Transfer files to magefiles directory
Using a separate directory for the magefiles allows the go.mod for the
module more accurately reflect what the _module_ code uses, distinct
from what the _build system_ uses. The dependencies of this module, for
example, no longer include Revive and all its dependencies, so our
go.mod is much shorter.

Making this change required adding some more complexity to the
tool-installing tasks because they now rely on the go.mod of the
magefile directory, not the go.mod of the main module, and there doesn't
seem to be a good way to discover that file's location programmatically.
Instead, a new function is added to the installation tasks to explicitly
tell where the mage go.mod file is.