Build the builder in the magefile directory
The `go run` command uses go.mod from the current directory, not the
directory where the mage.go source file is. That happens to work in
_this_ project because the main module and the magefiles module _both_
use mage. In other projects, that wouldn't work. To set a good example,
we explicitly build the command in the subdirectory and then run it from
the directory where we want the main build to happen.