Allow running tests with Ginkgo
Build and test dependencies now support a `UseGinkgo` function that
change them to use the ginkgo command to build and run test binaries
instead of using `go test`.

Making this change involved publishing several of the types that the
dependency functions return.

There is no longer a separate "Dep" version of the Build and Test
commands; they're now always the dependency version, but without the
"Dep" name.