Fix Package.HasTest, and test it
The logic in this function was inverted, somehow. Now it's fixed, and
there are new tests to confirm it. It's a little hokey because the
negative test requires having a package that has no tests, and there
just isn't a natural occasion in this module for such a condition, so we
add a special "notest" package just for that.