Specify module directory in Stringer
To installer the stringer utility, we need to know which version, and we
get the version from the go.mod file, so we need to know where that file
is. Mockgen probably needs this information as well, but I'm in the
process of redoing the mockgen support. If it's still needed afterward,
it'll come later.