Rewrite mockgen tasks
Mockgen is now configured with an external file. This allows Mage to
calculate dependencies more accurately, so we'll only generate mocks
when the relevant files have changed. The old way couldn't track changes
to the list of required mocks, so we'd either generate them on every
build, or never generate them.

There's a skeleton of an example that doesn't yet demonstrate anything
about the Mockgen target, and there are still some linting issues to
work through.